refactor: refatorando lógica para gerar nova questão.
